+++ cdk/backend/function/expressLambda/api/categories/queries.ts
