+++ cdk/backend/function/expressLambda/api/questionAnswers/router.ts
@@ -1,7 +1,5 @@
 import express from 'express'
-import QuestionAnswer, {
-  CreateQuestionAnswerProp,
-} from './questionAnswersQuery'
+import QuestionAnswer, { CreateQuestionAnswerProp } from './queries'
 
 const router = express.Router()
 
@@ -12,8 +10,7 @@ interface ReqParams {
 // List all questionAnswers
 router.get('/', async (req, res, next) => {
   try {
-    const listQuestionAnswerResponse =
-      await QuestionAnswer.listQuestionAnswers()
+    const listQuestionAnswerResponse = await QuestionAnswer.listQuestionAnswers()
     res.status(200).json(listQuestionAnswerResponse)
   } catch (err) {
     console.error(err)
@@ -41,8 +38,9 @@ router.post<unknown, unknown, CreateQuestionAnswerProp>(
   '/',
   async (req, res, next) => {
     try {
-      const createQuestionAnswerResponse =
-        await QuestionAnswer.createQuestionAnswer(req.body)
+      const createQuestionAnswerResponse = await QuestionAnswer.createQuestionAnswer(
+        req.body
+      )
       res.status(201).json(createQuestionAnswerResponse)
     } catch (err) {
       console.error(err)
@@ -57,8 +55,10 @@ router.patch<ReqParams, unknown, CreateQuestionAnswerProp>(
   async (req, res, next) => {
     try {
       const { questionAnswerId } = req.params
-      const updateQuestionAnswerResponse =
-        await QuestionAnswer.updateQuestionAnswer(questionAnswerId, req.body)
+      const updateQuestionAnswerResponse = await QuestionAnswer.updateQuestionAnswer(
+        questionAnswerId,
+        req.body
+      )
       res.status(200).json(updateQuestionAnswerResponse)
     } catch (err) {
       console.error(err)
@@ -71,8 +71,9 @@ router.patch<ReqParams, unknown, CreateQuestionAnswerProp>(
 router.delete<ReqParams>('/:questionAnswerId', async (req, res, next) => {
   try {
     const { questionAnswerId } = req.params
-    const deleteQuestionAnswerResponse =
-      await QuestionAnswer.deleteQuestionAnswer(questionAnswerId)
+    const deleteQuestionAnswerResponse = await QuestionAnswer.deleteQuestionAnswer(
+      questionAnswerId
+    )
     res.status(200).json(deleteQuestionAnswerResponse)
   } catch (err) {
     console.error(err)
@@ -92,9 +93,10 @@ router.post<unknown, unknown, CreateQuestionAnswerProp[]>(
     try {
       const responses: Response[] = []
       await Promise.all(
-        req.body.map(async (qa) => {
-          const createQuestionAnswerResponse: Response =
-            await QuestionAnswer.createQuestionAnswer(qa)
+        req.body.map(async qa => {
+          const createQuestionAnswerResponse: Response = await QuestionAnswer.createQuestionAnswer(
+            qa
+          )
           responses.push(createQuestionAnswerResponse)
         })
       )
